lint: [move-only] Move docs related lints to lint_docs.rs
Also, rename lint_doc to lint_doc_args.
